requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
numpy==1.24.3
tensorflow==2.13.0
opencv-python==4.8.1.78
scikit-learn==1.3.0
python-dotenv==1.0.0
pydantic==2.5.0
gunicorn==21.2.0
